requirement.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
djangorestframework==3.15.2
jwt==1.3.1
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
